+++ Assets/Scripts/Behaviours/StatusLineManager.cs
@@ -16,6 +16,8 @@ namespace Ventura.Behaviours
     public class StatusLineManager : MonoBehaviour
     {
         private static StatusLineManager instance;
+        public static StatusLineManager Instance { get { return instance; } }
+
         private TextMeshProUGUI _statusLine;
 
         void Start()
@@ -24,17 +26,14 @@ namespace Ventura.Behaviours
             _statusLine = GameObject.Find("Status Line").GetComponent<TextMeshProUGUI>();
         }
 
-        public static void DisplayStatus(string msg, StatusSeverity severity = StatusSeverity.Normal)
+        public void Clear()
         {
-            DebugUtils.Log("DisplayStatus");
-
-            if (instance != null)
-                instance.displayStatus(msg, severity);
+            _statusLine.text = "";
         }
 
-        private void displayStatus(string msg, StatusSeverity severity)
+        public void Display(string msg, StatusSeverity severity = StatusSeverity.Normal)
         {
-            DebugUtils.Log("displayStatus");
+            DebugUtils.Log("Display");
             DebugUtils.Log(msg);
             _statusLine.color = GraphicsConfig.StatusLineColors[severity];
             _statusLine.text = msg;