+++ Assets/Scripts/Behaviours/ViewManager.cs
@@ -1,22 +1,38 @@
 
 using System.Collections.Generic;
 using UnityEngine;
+using Ventura.GameLogic;
+using Ventura.Util;
 
 namespace Ventura.Behaviours
 {
+    public interface ModalUIManager
+    {
+        public void UpdateData();
+    }
+
 
     public class ViewManager : MonoBehaviour
     {
-        private const int foregroundDepth = 99;
-        private const int hiddenDepth = 0;
+        public GameObject inventoryUITemplate;
+        public GameObject skillsUITemplate;
 
         public enum ViewId
         {
             Map,
             Inventory,
+            Skills,
         }
 
+        private const int foregroundDepth = 99;
+        private const int hiddenDepth = -99;
+
+        private Dictionary<ViewId, GameObject> _viewTemplates;
+
         private Camera _modalUICamera;
+        private Transform _modalUIRoot;
+        //private Dictionary<ViewId, GameObject> _modalUIObjs = new();
+
         private ViewId _currView;
         private Dictionary<ViewId, KeyboardInputReceiver> _keyboardInputReceivers = new();
 
@@ -26,9 +42,18 @@ namespace Ventura.Behaviours
         void Start()
         {
             _modalUICamera = GameObject.Find("Modal UI Camera").GetComponent<Camera>();
+            _modalUIRoot = GameObject.Find("Modal UI Canvas").transform.Find("Instantiation Target");
 
             _keyboardInputReceivers.Add(ViewId.Map, new MapInputReceiver(this));
             _keyboardInputReceivers.Add(ViewId.Inventory, new InventoryInputReceiver(this));
+            _keyboardInputReceivers.Add(ViewId.Skills, new SkillsInputReceiver(this));
+
+
+            _viewTemplates = new()
+            {
+                { ViewId.Inventory, inventoryUITemplate },
+                { ViewId.Skills,    skillsUITemplate    },
+            };
 
             _currView = ViewId.Map;
         }
@@ -43,9 +68,32 @@ namespace Ventura.Behaviours
                 return;
 
             if (targetView == ViewId.Map)
+            {
                 _modalUICamera.depth = hiddenDepth;
+            }
             else
+            {
                 _modalUICamera.depth = foregroundDepth;
+                //TODO: reuse old _modalUIObjs
+
+                //GameObject modalUIObj;
+                //TODO: lastModalUIObj.SetActive(false);
+                //if (_modalUIObjs.ContainsKey(targetView))
+                //    modalUIObj = _modalUIObjs[targetView];
+
+                //else {
+                //    modalUIObj = Instantiate(inventoryUITemplate);
+                //    modalUIObj.transform.SetParent(_modalUIRoot);
+                //}
+                //_modalUIObjs[targetView] = modalUIObj;
+                //modalUIObj.GetComponent<ModalUIManager>().UpdateData(); //too early to call it here anyway
+
+
+                UnityUtils.RemoveAllChildren(_modalUIRoot);
+
+                var modalUIObj = Instantiate(_viewTemplates[targetView]);
+                modalUIObj.transform.SetParent(_modalUIRoot, false);
+            }
 
             _currView = targetView;
         }