+++ Assets/Scripts/Util/DebugUtils.cs
@@ -9,5 +9,10 @@ namespace Ventura.Util
             UnityEngine.Debug.Log(msg);
         }
 
+        public static void Error(string msg)
+        {
+            UnityEngine.Debug.LogError(msg);
+        }
+
     }
 }